+++ server/src/main/java/com/inventage/portal/gateway/proxy/config/dynamic/DynamicConfiguration.java
@@ -1,21 +1,5 @@
 package com.inventage.portal.gateway.proxy.config.dynamic;
 
-import java.net.MalformedURLException;
-import java.net.URISyntaxException;
-import java.net.URL;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Base64;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Map.Entry;
-import java.util.Set;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import io.vertx.core.CompositeFuture;
 import io.vertx.core.Future;
 import io.vertx.core.Promise;
@@ -28,6 +12,14 @@ import io.vertx.json.schema.SchemaRouter;
 import io.vertx.json.schema.SchemaRouterOptions;
 import io.vertx.json.schema.common.dsl.ObjectSchemaBuilder;
 import io.vertx.json.schema.common.dsl.Schemas;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.net.MalformedURLException;
+import java.net.URISyntaxException;
+import java.net.URL;
+import java.util.*;
+import java.util.Map.Entry;
 
 /**
  * It defines the structure of the dynamic configuration.
@@ -87,9 +79,12 @@ public class DynamicConfiguration {
     public static final String MIDDLEWARE_SESSION_BAG_WHITHELISTED_COOKIE_NAME = "name";
     public static final String MIDDLEWARE_SESSION_BAG_WHITHELISTED_COOKIE_PATH = "path";
 
+    public static final String MIDDLEWARE_CONTROL_API = "controlApi";
+    public static final String MIDDLEWARE_CONTROL_API_ACTION = "action";
+
     public static final List<String> MIDDLEWARE_TYPES = Arrays.asList(MIDDLEWARE_REPLACE_PATH_REGEX,
             MIDDLEWARE_REDIRECT_REGEX, MIDDLEWARE_HEADERS, MIDDLEWARE_AUTHORIZATION_BEARER, MIDDLEWARE_BEARER_ONLY,
-            MIDDLEWARE_OAUTH2, MIDDLEWARE_SHOW_SESSION_CONTENT, MIDDLEWARE_SESSION_BAG);
+            MIDDLEWARE_OAUTH2, MIDDLEWARE_SHOW_SESSION_CONTENT, MIDDLEWARE_SESSION_BAG, MIDDLEWARE_CONTROL_API);
 
     public static final String SERVICES = "services";
     public static final String SERVICE_NAME = "name";
@@ -124,6 +119,7 @@ public class DynamicConfiguration {
                 .property(MIDDLEWARE_HEADERS_REQUEST, Schemas.objectSchema())
                 .property(MIDDLEWARE_HEADERS_RESPONSE, Schemas.objectSchema())
                 .property(MIDDLEWARE_SESSION_BAG_WHITHELISTED_COOKIES, Schemas.arraySchema())
+                .property(MIDDLEWARE_CONTROL_API_ACTION, Schemas.stringSchema())
                 .allowAdditionalProperties(false);
 
         ObjectSchemaBuilder middlewareSchema = Schemas.objectSchema()
@@ -590,6 +586,15 @@ public class DynamicConfiguration {
                     }
                     break;
                 }
+                case MIDDLEWARE_CONTROL_API:
+                    String action = mwOptions.getString(MIDDLEWARE_CONTROL_API_ACTION);
+                    if (action == null) {
+                        return Future.failedFuture(String.format("%s: No control api action defined [SESSION_TRERMINATE]", mwType));
+                    }
+                    if (!Objects.equals(action, "SESSION_TERMINATE")) {
+                        return Future.failedFuture(String.format("%s: Not supported control api action defined [SESSION_TRERMINATE]", mwType));
+                    }
+                    break;
                 default: {
                     return Future.failedFuture(String.format("Unknown middleware: '%s'", mwType));
                 }