+++ server/src/main/java/com/inventage/portal/gateway/proxy/middleware/MiddlewareFactory.java
@@ -1,15 +1,14 @@
 package com.inventage.portal.gateway.proxy.middleware;
 
-import java.util.Optional;
-import java.util.ServiceLoader;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import io.vertx.core.Future;
 import io.vertx.core.Vertx;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.web.Router;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.Optional;
+import java.util.ServiceLoader;
 
 /**
  * Service interface for providing middlewares. Implementations must add an entry with the fully
@@ -30,10 +29,7 @@ public interface MiddlewareFactory {
             final Optional<MiddlewareFactory> middleware = ServiceLoader.load(MiddlewareFactory.class).stream()
                     .map(ServiceLoader.Provider::get).filter(instance -> instance.provides().equals(middlewareName))
                     .findFirst();
-            if (middleware.isPresent()) {
-                return middleware.get();
-            }
-            return null;
+            return middleware.orElse(null);
         }
     }